Revert "Test/parser"
