+++ ast/walk.go
@@ -1,7 +1,5 @@
 package ast
 
-import "reflect"
-
 type Visitor interface {
 	Visit(node Node) (w Visitor)
 }
@@ -31,19 +29,6 @@ func Walk(v Visitor, node Node) {
 	case *Text:
 	case *Ident:
 	default:
-		panic("cannot walk node of type: " + reflect.TypeOf(n).String())
+		return
 	}
-}
-
-type inspector func(Node) bool
-
-func (f inspector) Visit(node Node) Visitor {
-	if f(node) {
-		return f
-	}
-	return nil
-}
-
-func Inspect(node Node, f func(Node) bool) {
-	Walk(inspector(f), node)
 }
