+++ lexer/lexer.go
@@ -2,7 +2,7 @@ package lexer
 
 import (
 	"fmt"
-	source "go/token"
+	gtoken "go/token"
 	"strconv"
 	"strings"
 	"unicode/utf8"
@@ -18,7 +18,7 @@ const (
 type stateFunc func(*Lexer) stateFunc
 
 type Lexer struct {
-	file   *source.File
+	file   *gtoken.File
 	input  string
 	tokens chan token.Token
 
@@ -29,20 +29,15 @@ type Lexer struct {
 	width     int
 }
 
-func NewLexer(file *source.File, input string) *Lexer {
+func NewLexer(file *gtoken.File, input string) *Lexer {
 	l := &Lexer{
 		input: input,
 		// Channel must be large enough to support the largest
 		// amount of tokens that can be outputed from a single state
 		tokens: make(chan token.Token, 6),
-		state:  LexText,
+		state:  lexText,
 		file:   file,
 	}
-	return l
-}
-
-func (l *Lexer) WithState(state stateFunc) *Lexer {
-	l.state = state
 	return l
 }
 
@@ -67,11 +62,10 @@ func (l *Lexer) emit(typ token.TokenType) {
 
 	assert.Assert(l.pos > l.start, "pos must be past start")
 
-	literal := l.input[l.start:l.pos]
 	tok := token.Token{
-		Pos:     l.file.Pos(l.pos - len(literal)),
+		Pos:     l.file.Pos(l.pos),
 		Type:    typ,
-		Literal: literal,
+		Literal: l.input[l.start:l.pos],
 	}
 	l.tokens <- tok
 	l.start = l.pos
@@ -165,7 +159,7 @@ func (l *Lexer) discard() {
 	l.start = l.pos
 }
 
-func LexText(l *Lexer) stateFunc {
+func lexText(l *Lexer) stateFunc {
 	assert.AssertNotNil(l)
 
 	l.skipWhitespace()
@@ -182,10 +176,10 @@ func LexText(l *Lexer) stateFunc {
 	if l.pos > l.start {
 		l.emit(token.TEXT)
 	}
-	return LexTag
+	return lexTag
 }
 
-func LexTag(l *Lexer) stateFunc {
+func lexTag(l *Lexer) stateFunc {
 	assert.AssertNotNil(l)
 
 	ch := l.next()
@@ -196,7 +190,7 @@ func LexTag(l *Lexer) stateFunc {
 		l.emit(token.SLASH)
 	}
 
-	l.runUntil(" />")
+	l.runUntil("> ")
 	if l.peek() != eof {
 		l.emit(token.IDENT)
 	} else {
@@ -211,19 +205,15 @@ func LexTag(l *Lexer) stateFunc {
 	// todo: check all types of empty characters
 	l.acceptRun(" ")
 
-	if l.accept("/") {
-		l.emit(token.SLASH)
-	}
-
 	if l.accept(">") {
 		l.emit(token.RIGHT_CHEVRON)
-		return LexText
+		return lexText
 	}
 
-	return LexAttribute
+	return lexAttribute
 }
 
-func LexAttribute(l *Lexer) stateFunc {
+func lexAttribute(l *Lexer) stateFunc {
 	assert.Assert(!l.accept(" "), "expected no empty characters")
 
 	l.skipWhitespace()
@@ -272,8 +262,8 @@ func LexAttribute(l *Lexer) stateFunc {
 
 	if l.accept(">") {
 		l.emit(token.RIGHT_CHEVRON)
-		return LexText
+		return lexText
 	}
 
-	return LexAttribute
+	return lexAttribute
 }
