+++ parser/parser.go
@@ -1,12 +1,7 @@
 package parser
 
 import (
-	"bytes"
-	"errors"
 	"fmt"
-	source "go/token"
-	"io"
-	"os"
 	"strings"
 
 	"github.com/tifye/flamingo/assert"
@@ -14,61 +9,6 @@ import (
 	"github.com/tifye/flamingo/lexer"
 	"github.com/tifye/flamingo/token"
 )
-
-func ParseElement(src any) (*ast.Element, error) {
-	input, err := readSource("", src)
-	if err != nil {
-		return nil, err
-	}
-
-	fset := source.NewFileSet()
-	file := fset.AddFile("", fset.Base(), len(input))
-	l := lexer.NewLexer(file, string(input)).WithState(lexer.LexTag)
-	p := NewParser(l)
-	el := p.parseElement()
-	if n := len(p.Errors()); n > 0 {
-		return el, errors.New(strings.Join(p.Errors(), "; "))
-	}
-	return el, nil
-}
-
-func ParseFile(fset *source.FileSet, filename string, src any) (*ast.File, error) {
-	input, err := readSource(filename, src)
-	if err != nil {
-		return nil, fmt.Errorf("reading source: %s", err)
-	}
-
-	file := fset.AddFile(filename, fset.Base(), len(input))
-	l := lexer.NewLexer(file, string(input))
-	p := NewParser(l)
-
-	fileNode := p.Parse()
-	if len(p.errors) == 0 {
-		return fileNode, nil
-	}
-	return fileNode, fmt.Errorf("%v", p.errors)
-}
-
-func readSource(filename string, src any) ([]byte, error) {
-	if src != nil {
-		switch s := src.(type) {
-		case string:
-			return []byte(s), nil
-		case []byte:
-			return s, nil
-		case *bytes.Buffer:
-			if s != nil {
-				return s.Bytes(), nil
-			}
-		case io.Reader:
-			return io.ReadAll(s)
-		}
-
-		return nil, errors.New("invalid source type")
-	}
-
-	return os.ReadFile(filename)
-}
 
 type Parser struct {
 	l         *lexer.Lexer
@@ -148,7 +88,7 @@ func (p *Parser) parseElement() (el *ast.Element) {
 		return nil
 	}
 
-	element := &ast.Element{
+	comp := &ast.Element{
 		LeftChevron: p.curToken.Pos - 1,
 		Name: &ast.Ident{
 			Position: p.curToken.Pos,
@@ -159,27 +99,18 @@ func (p *Parser) parseElement() (el *ast.Element) {
 	}
 	defer func() {
 		if el != nil {
-			assert.Assert(element.RightChevron.IsValid(), "expected right chevron location to be set")
+			assert.Assert(comp.RightChevron.IsValid(), "expected right chevron location to be set")
 		}
 	}()
 
 	for p.tryPeek(token.IDENT) {
 		attr := p.parseAttribute()
 		if attr != nil {
-			element.Attrs = append(element.Attrs, attr)
+			comp.Attrs = append(comp.Attrs, attr)
 		}
 	}
 
-	if p.tryPeek(token.SLASH) {
-		if !p.expectPeek(token.RIGHT_CHEVRON) {
-			return nil
-		}
-
-		element.RightChevron = p.curToken.Pos
-		return element
-	}
-
-	assert.Assert(p.isPeekToken(token.RIGHT_CHEVRON), "expected next token to be a right chevron but got ")
+	assert.Assert(p.isPeekToken(token.RIGHT_CHEVRON), "expected next token to be a right chevron")
 	_ = p.expectPeek(token.RIGHT_CHEVRON)
 
 	for {
@@ -188,7 +119,7 @@ func (p *Parser) parseElement() (el *ast.Element) {
 		if el == nil {
 			break
 		}
-		element.Nodes = append(element.Nodes, el)
+		comp.Nodes = append(comp.Nodes, el)
 	}
 
 	if !p.expectPeek(token.SLASH) {
@@ -199,8 +130,8 @@ func (p *Parser) parseElement() (el *ast.Element) {
 		return nil
 	}
 
-	if p.curToken.Literal != element.Name.Name {
-		p.errorf("unexpected closing tag %s, expected %s", p.curToken.Literal, element.Name.Name)
+	if p.curToken.Literal != comp.Name.Name {
+		p.errorf("unexpected closing tag %s, expected %s", p.curToken.Literal, comp.Name.Name)
 		return nil
 	}
 
@@ -208,9 +139,9 @@ func (p *Parser) parseElement() (el *ast.Element) {
 		return nil
 	}
 
-	element.RightChevron = p.curToken.Pos
+	comp.RightChevron = p.curToken.Pos
 
-	return element
+	return comp
 }
 
 func (p *Parser) parseAttribute() *ast.Attribute {
