+++ parser/parser_test.go
@@ -1,19 +1,17 @@
 package parser
 
 import (
-	source "go/token"
-	"slices"
+	gtoken "go/token"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"github.com/tifye/flamingo/ast"
 	"github.com/tifye/flamingo/lexer"
 )
 
 func TestNestedSimpleComponents(t *testing.T) {
 	input := `<div class="bg-rose-500">mino</div>`
-	fset := source.NewFileSet()
+	fset := gtoken.NewFileSet()
 	f := fset.AddFile("", fset.Base(), len(input))
 	l := lexer.NewLexer(f, input)
 	p := NewParser(l)
@@ -24,73 +22,6 @@ func TestNestedSimpleComponents(t *testing.T) {
 	require.NotNil(t, root)
 	require.NotNil(t, root.Fragment, "expected a Fragment node")
 	assert.Equal(t, 1, len(root.Fragment.Nodes), "expected component to contain 1 node")
-}
-
-func TestParseElement(t *testing.T) {
-	t.Run("<div></div>", func(t *testing.T) {
-		input := `<div></div>`
-		el, err := ParseElement(input)
-		assert.NoError(t, err)
-		assert.NotNil(t, el)
-		ast.Inspect(el, func(node ast.Node) bool {
-			switch n := node.(type) {
-			case *ast.Element:
-				assert.Equal(t, "div", n.Name.Name)
-				assert.Equal(t, 1, int(n.LeftChevron))
-				assert.Equal(t, 11, int(n.RightChevron))
-			}
-			return true
-		})
-	})
-
-	t.Run("<div><meep><div></div></meep><mino></mino></div>", func(t *testing.T) {
-		input := `<div><meep><div></div></meep><mino></mino></div>`
-		elements := []string{"div", "div", "meep", "mino"}
-		el, err := ParseElement(input)
-		assert.NoError(t, err)
-		assert.NotNil(t, el)
-		ast.Inspect(el, func(node ast.Node) bool {
-			switch n := node.(type) {
-			case *ast.Element:
-				idx := slices.Index(elements, n.Name.Name)
-				elements = slices.Delete(elements, idx, idx+1)
-			}
-			return true
-		})
-		assert.Empty(t, elements)
-	})
-
-	t.Run("<self />", func(t *testing.T) {
-		input := `<self />`
-		el, err := ParseElement(input)
-		assert.NoError(t, err)
-		assert.NotNil(t, el)
-		ast.Inspect(el, func(node ast.Node) bool {
-			switch n := node.(type) {
-			case *ast.Element:
-				assert.Equal(t, "self", n.Name.Name)
-				assert.Equal(t, 1, int(n.LeftChevron))
-				assert.Equal(t, 8, int(n.RightChevron))
-			}
-			return true
-		})
-	})
-}
-
-func TestAttribute(t *testing.T) {
-	t.Run(`empty string literal`, func(t *testing.T) {
-		input := `<test isTrue="" />`
-		el, err := ParseElement(input)
-		assert.NoError(t, err)
-		assert.NotNil(t, el)
-	})
-
-	t.Run(`boolean attribute: <test isTrue />`, func(t *testing.T) {
-		input := `<test isTrue />`
-		el, err := ParseElement(input)
-		assert.NoError(t, err)
-		assert.NotNil(t, el)
-	})
 }
 
 func noParserErrors(t *testing.T, p *Parser) {
